Release v0.8.x (#208)
* serialize ts as FloatField

* Bump version: 0.8.15 → 0.8.16

* 0.8.16 client codegen ✨

* Resolves TypeError

 File "/app/print_nanny_webapp/telemetry/subscribers/octoprint_events.py", line 229, in on_octoprint_event
    data = dict(
TypeError: type object got multiple values for keyword argument
'printer_state'

* indicate deleted is a read-only field

* Bump version: 0.8.16 → 0.8.17

* 0.8.17 client codegen ✨

* do not serialize deleted field at all

* attempt #2 at excluding deleted field

* Bump version: 0.8.17 → 0.8.18

* 0.8.18 client codegen ✨

+++ clients/kotlin/src/main/kotlin/com/print-nanny/client/models/OctoPrintEventRequest.kt
@@ -50,7 +50,7 @@ data class OctoPrintEventRequest (
     @Json(name = "octoprint_device")
     val octoprintDevice: kotlin.Int,
     @Json(name = "ts")
-    val ts: kotlin.Int? = null,
+    val ts: kotlin.Float? = null,
     @Json(name = "event_data")
     val eventData: kotlin.collections.Map<kotlin.String, AnyType>? = null,
     @Json(name = "temperature")
