+++ clients/kotlin/src/main/kotlin/com/print-nanny/client/models/PrinterEventRequest.kt
@@ -54,7 +54,7 @@ data class PrinterEventRequest (
     @Json(name = "octoprint_device")
     val octoprintDevice: kotlin.Int,
     @Json(name = "ts")
-    val ts: kotlin.Int? = null,
+    val ts: kotlin.Float? = null,
     @Json(name = "event_data")
     val eventData: kotlin.collections.Map<kotlin.String, AnyType>? = null,
     @Json(name = "temperature")
