+++ clients/python/print_nanny_client/models/print_nanny_plugin_event.py
@@ -37,7 +37,7 @@ class PrintNannyPluginEvent(object):
     """
     openapi_types = {
         'id': 'int',
-        'ts': 'int',
+        'ts': 'float',
         'event_type': 'PrintNannyPluginEventEventTypeEnum',
         'octoprint_environment': 'OctoprintEnvironment',
         'octoprint_printer_data': 'OctoprintPrinterData',
@@ -141,7 +141,7 @@ class PrintNannyPluginEvent(object):
 
 
         :return: The ts of this PrintNannyPluginEvent.  # noqa: E501
-        :rtype: int
+        :rtype: float
         """
         return self._ts
 
@@ -151,7 +151,7 @@ class PrintNannyPluginEvent(object):
 
 
         :param ts: The ts of this PrintNannyPluginEvent.  # noqa: E501
-        :type ts: int
+        :type ts: float
         """
 
         self._ts = ts
