+++ clients/python/print_nanny_client/models/telemetry_event.py
@@ -37,7 +37,7 @@ class TelemetryEvent(object):
     """
     openapi_types = {
         'id': 'int',
-        'ts': 'int',
+        'ts': 'float',
         'event_type': 'TelemetryEventEventTypeEnum',
         'octoprint_environment': 'OctoprintEnvironment',
         'octoprint_printer_data': 'OctoprintPrinterData',
@@ -141,7 +141,7 @@ class TelemetryEvent(object):
 
 
         :return: The ts of this TelemetryEvent.  # noqa: E501
-        :rtype: int
+        :rtype: float
         """
         return self._ts
 
@@ -151,7 +151,7 @@ class TelemetryEvent(object):
 
 
         :param ts: The ts of this TelemetryEvent.  # noqa: E501
-        :type ts: int
+        :type ts: float
         """
 
         self._ts = ts
