+++ clients/python/print_nanny_client/models/telemetry_event_polymorphic.py
@@ -37,7 +37,7 @@ class TelemetryEventPolymorphic(object):
     """
     openapi_types = {
         'id': 'int',
-        'ts': 'int',
+        'ts': 'float',
         'event_type': 'PrintNannyPluginEventEventTypeEnum',
         'octoprint_environment': 'OctoprintEnvironment',
         'octoprint_printer_data': 'OctoprintPrinterData',
@@ -148,7 +148,7 @@ class TelemetryEventPolymorphic(object):
 
 
         :return: The ts of this TelemetryEventPolymorphic.  # noqa: E501
-        :rtype: int
+        :rtype: float
         """
         return self._ts
 
@@ -158,7 +158,7 @@ class TelemetryEventPolymorphic(object):
 
 
         :param ts: The ts of this TelemetryEventPolymorphic.  # noqa: E501
-        :type ts: int
+        :type ts: float
         """
 
         self._ts = ts
