+++ print_nanny_webapp/telemetry/api/serializers.py
@@ -108,7 +108,7 @@ class OctoprintPrinterDataSerializer(serializers.Serializer):
 
 class TelemetryEventSerializer(serializers.ModelSerializer):
 
-    ts = serializers.IntegerField(required=False)
+    ts = serializers.FloatField(required=False)
     event_type = serializers.ChoiceField(choices=TelemetryEventType.choices)
     octoprint_environment = OctoprintEnvironmentSerializer()
     octoprint_printer_data = OctoprintPrinterDataSerializer()
